+++ src/main/adapters/express-route-adapter.ts
@@ -1,4 +1,4 @@
-import { Controller, HttpRequest } from '../../../presentation/protocols'
+import { Controller, HttpRequest } from '../../presentation/protocols'
 import { Request, RequestHandler, Response } from 'express'
 
 export const adaptRoute = (controller: Controller): RequestHandler => {
