+++ src/main/factories/controllers/login/login/login-controller-factory.ts
@@ -1,0 +1,10 @@
+import { LoginController } from '../../../../../presentation/controllers/login/login/login-controller'
+import { Controller } from '../../../../../presentation/protocols'
+import { makeLogControllerDecorator } from '../../../decorators/log-controller-decorator-factory'
+import { makeDbAuthentication } from '../../../usecases/account/authentication/db-authentication-factory'
+import { makeLoginValidation } from './login-validation-factory'
+
+export const makeLoginController = (): Controller => {
+  const controller = new LoginController(makeDbAuthentication(), makeLoginValidation())
+  return makeLogControllerDecorator(controller)
+}
