--- src/main/factories/usecases/add-account/db-add-account-factory.ts
@@ -1,11 +1,0 @@
-import { BcryptAdapter } from '../../../../infra/criptography/bcrypt-adapter/bcrypt-adapter'
-import { AccountMongoRepository } from '../../../../infra/db/mongodb/account/account-mongo-repository'
-import { AddAccount } from '../../../../domain/usecases/add-account'
-import { DbAddAccount } from '../../../../data/usecases/add-account/db-add-account'
-
-export const makeDbAddAccount = (): AddAccount => {
-  const salt = 12
-  const bcryptAdapter = new BcryptAdapter(salt)
-  const accountMongoRepository = new AccountMongoRepository()
-  return new DbAddAccount(bcryptAdapter, accountMongoRepository, accountMongoRepository)
-}
