+++ src/main/factories/usecases/survey/add-survey/db-add-survey-factory.ts
@@ -1,0 +1,8 @@
+import { AddSurvey } from '../../../../../domain/usecases/add-survey'
+import { SurveyMongoRepository } from '../../../../../infra/db/mongodb/survey/survey-mongo-repository'
+import { DbAddSurvey } from '../../../../../data/usecases/add-survey/db-add-survey'
+
+export const makeDbAddSurvey = (): AddSurvey => {
+  const surveyMongoRepository = new SurveyMongoRepository()
+  return new DbAddSurvey(surveyMongoRepository)
+}
