+++ src/main/routes/login-routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
-import { adaptRoute } from '../adapters/express/express-route-adapter'
-import { makeSignUpController } from '../factories/controllers/signup/signup-controller-factory'
-import { makeLoginController } from '../factories/controllers/login/login-controller-factory'
+import { adaptRoute } from '../adapters/express-route-adapter'
+import { makeSignUpController } from '../factories/controllers/login/signup/signup-controller-factory'
+import { makeLoginController } from '../factories/controllers/login/login/login-controller-factory'
 
 export default (router: Router): void => {
   router.post('/signup', adaptRoute(makeSignUpController()))
