+++ src/main/routes/survey-routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
-import { adaptRoute } from '../adapters/express/express-route-adapter'
-import { makeAddSurveyController } from '../factories/controllers/add-survey/add-survey-controller-factory'
+import { adaptRoute } from '../adapters/express-route-adapter'
+import { makeAddSurveyController } from '../factories/controllers/survey/add-survey/add-survey-controller-factory'
 
 export default (router: Router): void => {
   router.post('/surveys', adaptRoute(makeAddSurveyController()))
